Refactor integration tests
Move dedicated integration tests to separate files
