+++ tests/integration/agents/supervisor/test_finalizer.py
@@ -1,242 +1,9 @@
-from collections.abc import Sequence
-
 import pytest
-from deepeval import assert_test, evaluate
-from deepeval.metrics import AnswerRelevancyMetric, GEval
-from deepeval.test_case import LLMTestCase, LLMTestCaseParams
-from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage
+from deepeval import evaluate
+from deepeval.test_case import LLMTestCase
+from langchain_core.messages import AIMessage, HumanMessage, SystemMessage
 
-from agents.common.constants import COMMON
-from agents.common.state import CompanionState, SubTask, UserInput
-from agents.k8s.agent import K8S_AGENT
-from agents.kyma.agent import KYMA_AGENT
-
-
-@pytest.fixture
-def answer_relevancy_metric(evaluator_model):
-    return AnswerRelevancyMetric(
-        threshold=0.6, model=evaluator_model, include_reason=True
-    )
-
-
-@pytest.fixture
-def semantic_similarity_metric(evaluator_model):
-    return GEval(
-        name="Semantic Similarity",
-        evaluation_steps=[
-            "Evaluate whether two answers are semantically similar or convey the same meaning.",
-            "Lightly penalize omissions of detail, focusing on the main idea",
-            "Vague language are permissible",
-        ],
-        evaluation_params=[
-            LLMTestCaseParams.ACTUAL_OUTPUT,
-            LLMTestCaseParams.EXPECTED_OUTPUT,
-        ],
-        model=evaluator_model,
-        threshold=0.7,
-    )
-
-
-# Correctness metric for not general queries that needs planning
-@pytest.fixture
-def planner_correctness_metric(evaluator_model):
-    return GEval(
-        name="Correctness",
-        criteria=""
-        "Determine whether the output is subtask(s) of the input and assigned to dedicated agent(s). "
-        "It is okay if the output contains one subtask."
-        "Check if the output is in valid JSON format"
-        "Verify that the JSON contains required keys: 'subtasks'",
-        evaluation_params=[
-            LLMTestCaseParams.ACTUAL_OUTPUT,
-            LLMTestCaseParams.EXPECTED_OUTPUT,
-        ],
-        model=evaluator_model,
-        threshold=0.8,
-    )
-
-
-def create_mock_state(messages: Sequence[BaseMessage], subtasks=None) -> CompanionState:
-    """Create a mock langgraph state for tests."""
-    if subtasks is None:
-        subtasks = []
-    user_input = UserInput(
-        query=messages[-1].content,
-        resource_kind=None,
-        resource_api_version=None,
-        resource_name=None,
-        namespace=None,
-    )
-
-    return CompanionState(
-        input=user_input,
-        messages=messages,
-        next="",
-        subtasks=subtasks,
-        final_response="",
-        error=None,
-    )
-
-
-@pytest.mark.parametrize(
-    "messages, expected_answer, general_query",
-    [
-        (
-            [
-                # tests if a general query is immediately answered by the planner
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="What is the capital of Germany?"),
-            ],
-            '{"response": "Berlin"}',
-            True,
-        ),
-        (
-            # tests if a Kyma related query is assigned to the Kyma agent
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="What is Kyma?"),
-            ],
-            '{"subtasks": [{"description": "What is Kyma?", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a Kyma related query is assigned to the Kyma agent
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="What is Kyma API Rule?"),
-            ],
-            '{"subtasks": [{"description": "What is Kyma API Rule?", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a Kubernetes related query is assigned to the Kubernetes agent
-            [
-                AIMessage(
-                    content="The `nginx` container in the `nginx-5dbddc77dd-t5fm2` pod is experiencing a "
-                    "`CrashLoopBackOff` state. The last termination reason was `StartError`"
-                    " with the message indicating a failure to create the containerd task "
-                    "due to a context cancellation."
-                ),
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="why the pod is failing?"),
-            ],
-            "{'subtasks': [{'description': 'why the pod is failing?', 'assigned_to': 'KubernetesAgent'}]}",
-            False,
-        ),
-        (
-            # tests if a Kubernetes related query is assigned to the Kubernetes agent
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="what is the status of my cluster?"),
-            ],
-            '{"subtasks": [{"description": "what is the status of my cluster?", "assigned_to": "KubernetesAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a query related to Kyma and Kubernetes is divided into the correct subtasks for both agents
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="What is Kubernetes? Explain Kyma function"),
-            ],
-            '{"subtasks": [{"description": "What is Kubernetes?", "assigned_to": "KubernetesAgent"},'
-            '{"description": "Explain Kyma function", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a query related to Kyma and Common is divided into the correct subtasks for both agents
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(
-                    content="Create a hello world app and deploy it with Kyma?"
-                ),
-            ],
-            '{"subtasks": [{"description": "Create a hello world app", "assigned_to": "Common"},'
-            '{"description": "deploy it with Kyma", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a query related to Kyma and Common is divided into the correct subtasks for both agents
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(
-                    content="Create a hello world app with python and deploy it with Kyma?"
-                ),
-            ],
-            '{"subtasks": [{"description": "Create a hello world app with python", "assigned_to": "Common"},'
-            '{"description": "deploy it with Kyma", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-        (
-            # tests if a complex query related to Kyma is divided correctly into two subtasks for the Kyma agent
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(
-                    content="how to enable eventing module and create a subscription for my app?"
-                ),
-            ],
-            '{"subtasks": [{"description": "How to enable eventing module?", "assigned_to": "KymaAgent"},'
-            '{"description": "How to create a subscription for my app?", "assigned_to": "KymaAgent"}]}',
-            False,
-        ),
-    ],
-)
-def test_invoke_planner(
-    messages,
-    expected_answer,
-    general_query,
-    companion_graph,
-    planner_correctness_metric,
-    answer_relevancy_metric,
-):
-    """Tests the invoke_planner method of SupervisorAgent."""
-    state = create_mock_state(messages)
-    result = companion_graph.supervisor_agent._invoke_planner(state)
-
-    test_case = LLMTestCase(
-        input=messages[-1].content,
-        actual_output=result.content,
-        expected_output=expected_answer,
-    )
-
-    if not general_query:
-        planner_correctness_metric.measure(test_case)
-
-        print(f"Score: {planner_correctness_metric.score}")
-        print(f"Reason: {planner_correctness_metric.reason}")
-
-        # Parse the output to check if it is in valid JSON format
-        companion_graph.plan_parser.parse(result.content)
-
-        assert_test(test_case, [planner_correctness_metric])
-    else:
-        assert_test(test_case, [answer_relevancy_metric])
+from integration.agents.test_common_node import create_mock_state
 
 
 @pytest.mark.parametrize(
@@ -586,205 +353,3 @@ def test_invoke_finalizer(
     assert all(
         result.success for result in eval_results.test_results
     ), "Not all metrics passed"
-
-
-@pytest.mark.parametrize(
-    "messages, subtasks, expected_answer",
-    [
-        # Single task:
-        # - Assigned K8sAgent is 'Next' agent
-        # - Assigned KymaAgent is 'Next' agent
-        # - Assigned Common is 'Next' agent
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=K8S_AGENT, status="pending"),
-            ],
-            K8S_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=KYMA_AGENT, status="pending"),
-            ],
-            KYMA_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=COMMON, status="pending"),
-            ],
-            COMMON,
-        ),
-        # Multiple tasks, all still pending:
-        # - First assigned K8sAgent is 'Next' agent
-        # - First assigned KymaAgent is 'Next' agent
-        # - First assigned Common is 'Next' agent
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=K8S_AGENT, status="pending"),
-                SubTask(description="Task 2", assigned_to=KYMA_AGENT, status="pending"),
-            ],
-            K8S_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=KYMA_AGENT, status="pending"),
-                SubTask(description="Task 2", assigned_to=K8S_AGENT, status="pending"),
-            ],
-            KYMA_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(description="Task 1", assigned_to=COMMON, status="pending"),
-                SubTask(description="Task 2", assigned_to=KYMA_AGENT, status="pending"),
-            ],
-            COMMON,
-        ),
-        # Multiple tasks, the first one is completed:
-        # - Second pending K8sAgent is 'Next' agent
-        # - Second pending KymaAgent is 'Next' agent
-        # - Second pending Common is 'Next' agent
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(
-                    description="Task 1", assigned_to=KYMA_AGENT, status="completed"
-                ),
-                SubTask(description="Task 2", assigned_to=K8S_AGENT, status="pending"),
-            ],
-            K8S_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(
-                    description="Task 1", assigned_to=K8S_AGENT, status="completed"
-                ),
-                SubTask(description="Task 2", assigned_to=KYMA_AGENT, status="pending"),
-            ],
-            KYMA_AGENT,
-        ),
-        (
-            [
-                SystemMessage(
-                    content="""
-                The user query is related to: {'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}
-                """
-                )
-            ],
-            [
-                SubTask(
-                    description="Task 1", assigned_to=KYMA_AGENT, status="completed"
-                ),
-                SubTask(description="Task 2", assigned_to=COMMON, status="pending"),
-            ],
-            COMMON,
-        ),
-    ],
-)
-def test_invoke_router(messages, subtasks, expected_answer, companion_graph):
-    """Tests the router method of CompanionGraph."""
-    state = create_mock_state(messages, subtasks)
-
-    result = companion_graph.supervisor_agent._route(state)
-
-    actual_output = result["next"]
-    expected_output = expected_answer
-
-    assert actual_output == expected_output
-
-
-@pytest.mark.parametrize(
-    "messages, expected_answer",
-    [
-        (
-            # tests that the Common node corretly answers a general non-technical query
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content="What is the capital of Germany?"),
-            ],
-            "Berlin",
-        ),
-        (
-            # tests that the Common node correctly answers a general programming related query
-            [
-                SystemMessage(
-                    content="The user query is related to: "
-                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
-                ),
-                HumanMessage(content='Write "Hello, World!" code in Python'),
-            ],
-            'Here is a simple "Hello, World!" program in Python: `print("Hello, World!")`',
-        ),
-    ],
-)
-def test_invoke_common_node(
-    messages, expected_answer, companion_graph, answer_relevancy_metric
-):
-    """Tests the invoke_common_node method of CompanionGraph."""
-    state = create_mock_state(messages)
-
-    result = companion_graph._invoke_common_node(state, messages[-1].content)
-
-    test_case = LLMTestCase(
-        input=messages[-1].content,
-        actual_output=result,
-        expected_output=expected_answer,
-    )
-
-    assert_test(test_case, [answer_relevancy_metric])
