+++ tests/integration/agents/test_common_node.py
@@ -1,0 +1,50 @@
+import pytest
+from deepeval import assert_test
+from deepeval.test_case import LLMTestCase
+from langchain_core.messages import HumanMessage, SystemMessage
+
+from integration.conftest import create_mock_state
+
+
+@pytest.mark.parametrize(
+    "messages, expected_answer",
+    [
+        (
+            # tests that the Common node corretly answers a general non-technical query
+            [
+                SystemMessage(
+                    content="The user query is related to: "
+                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
+                ),
+                HumanMessage(content="What is the capital of Germany?"),
+            ],
+            "Berlin",
+        ),
+        (
+            # tests that the Common node correctly answers a general programming related query
+            [
+                SystemMessage(
+                    content="The user query is related to: "
+                    "{'resource_api_version': 'v1', 'resource_namespace': 'nginx-oom'}"
+                ),
+                HumanMessage(content='Write "Hello, World!" code in Python'),
+            ],
+            'Here is a simple "Hello, World!" program in Python: `print("Hello, World!")`',
+        ),
+    ],
+)
+def test_invoke_common_node(
+    messages, expected_answer, companion_graph, answer_relevancy_metric
+):
+    """Tests the invoke_common_node method of CompanionGraph."""
+    state = create_mock_state(messages)
+
+    result = companion_graph._invoke_common_node(state, messages[-1].content)
+
+    test_case = LLMTestCase(
+        input=messages[-1].content,
+        actual_output=result,
+        expected_output=expected_answer,
+    )
+
+    assert_test(test_case, [answer_relevancy_metric])
